shell: add a way of listening to other widgets for enter/leave
Due to the many widgets in the panel, some enter and leave signals get
eaten up by some widgets here and there, so their ::enter-notify-event
should be connected to to make sure the cursor is still over the
panel. ::leave-notify-event isn't required as the panel will always
give us a reliable final leave event.

The reason why these signals are signalled or not signalled are not
fully understood.