gtk-shell: add hack to always keep app launcher onscreen
At the moment there's a problem where weston_view->output_mask for the
app launcher window isn't updated during the animation of moving it
onscreen. While offscreen ->output_mask is obviously 0 but during the
animation the correct output should be added to the mask.

To workaround this problem for now we'll keep the window always
onscreen but hidden by the panel so this problem is never hit. This is
a horrible hack and should be fixed properly as soon as possible.